Move shared DOM files to client
These are only used by client abstractions now. They're inlined in the Fizz code.
