+++ packages/react-dom-bindings/src/client/DOMNamespaces.js
