+++ packages/react-dom-bindings/src/client/HTMLNodeType.js
