+++ packages/react-dom-bindings/src/client/ReactDOMEventHandle.js
@@ -12,7 +12,7 @@ import type {ReactScopeInstance} from 'shared/ReactTypes';
 import type {
   ReactDOMEventHandle,
   ReactDOMEventHandleListener,
-} from '../shared/ReactDOMTypes';
+} from './ReactDOMEventHandleTypes';
 
 import {allNativeEvents} from '../events/EventRegistry';
 import {
@@ -21,7 +21,7 @@ import {
   doesTargetHaveEventHandle,
   addEventHandleToTarget,
 } from './ReactDOMComponentTree';
-import {ELEMENT_NODE} from '../shared/HTMLNodeType';
+import {ELEMENT_NODE} from './HTMLNodeType';
 import {listenToNativeEventForNonManagedEventTarget} from '../events/DOMPluginEventSystem';
 
 import {
