+++ packages/react-dom-bindings/src/client/ReactDOMEventHandleTypes.js
