+++ packages/react-dom-bindings/src/client/assertValidProps.js
