+++ packages/react-dom-bindings/src/client/getNodeForCharacterOffset.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-import {TEXT_NODE} from '../shared/HTMLNodeType';
+import {TEXT_NODE} from './HTMLNodeType';
 
 /**
  * Given any node return the first leaf node without children.
