+++ packages/react-dom-bindings/src/client/omittedCloseTags.js
