+++ packages/react-dom-bindings/src/events/plugins/SelectEventPlugin.js
@@ -23,7 +23,7 @@ import {registerTwoPhaseEvent} from '../EventRegistry';
 import getActiveElement from '../../client/getActiveElement';
 import {getNodeFromInstance} from '../../client/ReactDOMComponentTree';
 import {hasSelectionCapabilities} from '../../client/ReactInputSelection';
-import {DOCUMENT_NODE} from '../../shared/HTMLNodeType';
+import {DOCUMENT_NODE} from '../../client/HTMLNodeType';
 import {accumulateTwoPhaseListeners} from '../DOMPluginEventSystem';
 
 const skipSelectionChangeEvent =
